Block usage of vendor libs in hooks when NODE_ENV === 'test'

+++ src/models/picture/hooks.js
@@ -1,8 +1,12 @@
-const Aws     = require('../../vendor/aws');
-const config  = require('../../config');
+const Aws           = require('../../vendor/aws');
+const config        = require('../../config');
 
 module.exports = function(models, Picture) {
   Picture.hook('beforeCreate', (picture) => {
+    if ( config.NODE_ENV === 'test' ) {
+      return picture;
+    }
+
     return Aws.uploadPicture(picture)
       .then((url) => {
         return picture.url = url;
@@ -10,6 +14,10 @@ module.exports = function(models, Picture) {
   });
 
   Picture.hook('beforeDelete', (picture) => {
+    if ( config.NODE_ENV === 'test' ) {
+      return picture;
+    }
+
     return Aws.deleteFile(config.AWS_BUCKET_PICTURES, {
       Key: picture.id,
     });
